requirements.txt
langchain
langchain-groq
python_dotenv
langchian-community
langchain-openai
bs4
unstructured
pypdf
wikipedia
langchain-chroma
chromadb
faiss-cpu
langgraph
langsmith
langserve
